SDL mouse code now also works better
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@38442 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/sdl/vsyncarch.c
@@ -74,7 +74,7 @@ void vsyncarch_init(void)
 /* Sleep a number of timer units. */
 void vsyncarch_sleep(unsigned long delay)
 {
-    SDL_Delay(delay / VICE_SDL_TICKS_SCALE);
+    SDL_Delay((int)(delay / VICE_SDL_TICKS_SCALE));
 }
 
 void vsyncarch_presync(void)
